Implement sharded index artifacts and pipeline CI
