+++ docs/index.html
@@ -260,6 +260,7 @@
     <script src="https://cdn.jsdelivr.net/npm/fuse.js@7.0.0/dist/fuse.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js@2"></script>
     <script src="js/supabase-client.js"></script>
+    <script src="js/app-render.js"></script>
     <script src="js/app.js"></script>
 </body>
 </html>
