+++ tests/test_pipeline_contracts.py
@@ -40,6 +40,29 @@ def test_build_index_fails_closed_when_security_report_is_missing():
     assert "test -f docs/security-report.json" in workflow
 
 
+def test_build_index_runs_generated_size_guard_before_pages_upload():
+    workflow = read_repo_file(".github/workflows/build-index.yml")
+
+    guard_pos = workflow.index("scripts/check_generated_file_sizes.py")
+    upload_pos = workflow.index("actions/upload-pages-artifact")
+
+    assert guard_pos < upload_pos
+    assert "--include docs" in workflow
+
+
+def test_sync_data_runs_generated_size_guard_after_registry_rebuild():
+    workflow = read_repo_file(".github/workflows/sync-data.yml")
+
+    rebuild_pos = workflow.index("scripts/rebuild_registry.py")
+    guard_pos = workflow.index("scripts/check_generated_file_sizes.py")
+    commit_pos = workflow.index("Commit & push data repo changes")
+
+    assert rebuild_pos < guard_pos < commit_pos
+    assert "--include registry.json" in workflow
+    assert "--include registry-shards" in workflow
+    assert "--include docs" in workflow
+
+
 def test_metadata_compliance_refuses_unexpected_zero_target_scan():
     workflow = read_repo_file(".github/workflows/metadata-compliance.yml")
 
@@ -47,3 +70,13 @@ def test_metadata_compliance_refuses_unexpected_zero_target_scan():
     assert "metadata-advisory-zero-targets" in workflow
     assert "refusing to run metadata compliance with zero targets" in workflow
     assert "exit 1" in workflow
+
+
+def test_python_tests_workflow_runs_full_suite_with_coverage_gate():
+    workflow = read_repo_file(".github/workflows/python-tests.yml")
+
+    assert "name: Python Test Health" in workflow
+    assert "python -m pytest -q --cov-fail-under=50" in workflow
+    assert "--override-ini" not in workflow
+    assert "scripts/**" in workflow
+    assert "crawler/**" in workflow
